requirements.txt
# requirements.txt
kiteconnect==5.0.1
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
telethon==1.28.5
python-dotenv==1.0.0
schedule==1.2.0
ta-lib==0.4.26
yfinance==0.2.18
asyncio==3.4.3
websockets==11.0.3